frontend/requirements.txt
streamlit
requests
pillow
numpy
pandas
